HealthTrack_Ai/requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
mysql-connector-python==8.2.0
numpy==1.26.0
pandas==2.2.0
joblib==1.4.0
Werkzeug==3.0.0
python-dotenv==1.0.0